+++ lib/turntabler/event.rb
@@ -207,7 +207,12 @@ module Turntabler
     handle :song_started, :newsong do
       client.trigger(:song_ended) if room.current_song
       room.attributes = data['room']
-      room.current_song
+      current_song = room.current_song
+
+      # Update playlist order if it was played by the current user
+      current_song.dequeue if current_song.played_by == client.user
+
+      current_song
     end
 
     # The current song has ended